Do not allow disabling the main keyboard
It's not obvious that disabling it after installing another keyboard is
a very bad idea. Other keyboards rarely support Direct Boot and won't
work for entering the password, forcing recovery by plugging in a
physical keyboard. Even if they do support Direct Boot it's a bad idea
to rely on third party apps for basic functionality to keep working.

Change-Id: I7a79f6c9412eb87231e00b2fbed0684a0fede767