Privacy Guard - show WiFi scan as WiFi scan and not as Location
In cm-12.1, different from cm-13.0 and higher, the AppOps activity
'WiFi scan' is referred to as 'Location' in the app detail screen,
whilst is it correctly named 'WiFi scan' in the summary. Although
the WiFi scan indeed is suitable to be used to identify the location
(that is why Android 6 requires the location permission before
allowing getting the list of available WiFI networks), it should be
shown in privacy guard as the real activity, same as in higher releases.

In order to avoid missing language translations, the summary-string is
re-used as label for the app detail screen in cm_arrays.xml

Change-Id: I3f2a9b59eb14cd87dc52eb124e96c02805be8f88